add images in bito member landing pages
