+++ bito-world-frontend-main/src/App.css
@@ -1,1 +1,1 @@
- 
+ @import "tailwindcss";
