+++ bito-world-frontend-main/src/Components/MembershipLanding/components/Testimonials.jsx
@@ -1,39 +1,40 @@
 import Testimonial from "./Testimonial";
+import { testimonials } from "../../../utils/constants.js";
 
-const testimonials = [
-  {
-    program: "Platinum",
-    description:
-      "Being part of BITO means being part of Bihar’s growth and transformation. It’s about honoring our roots, and shaping a future. I’m grateful to contribute to this meaningful journey.",
-    image: new URL("../assets/testimonials/t1.jpg", import.meta.url),
-    name: "Farooque",
-    title: "Customer Success Manager, Dubai",
-  },
-  {
-    program: "Diamond",
-    description:
-      "BITO is more than an organization—it’s a movement. By empowering entrepreneurs, fostering innovation, and honoring Bihar’s rich heritage, it’s building bridges between our past and our global future. I’m proud to be part of this journey toward a more united and prosperous Bihar.",
-    image: new URL("../assets/testimonials/t2.jpg", import.meta.url),
-    name: "Upendra",
-    // title: "Product lead at EducationWave",
-  },
-  {
-    program: "Silver",
-    description:
-      "As an education leader, I admire BITO’s focus on experiential learning through trade exposure, internships, and global forums. These initiatives broaden horizons and make learning truly meaningful.",
-    image: new URL("../assets/testimonials/t3.jpg", import.meta.url),
-    name: "Abhilasha Singh",
-    title: "Principal Shining Star International School, Abu Dhabi",
-  },
-  {
-    program: "Gold",
-    description:
-      "We are expert in delivering exhibitions, interior and infrastructure services across worldwide having company presence in Patna, Dubai, Milano, Warsaw. Now just we launched in Patna Indl Area also.",
-    image: new URL("../assets/testimonials/t4.jpg", import.meta.url),
-    name: "Mukesh Pandey",
-    title: "Managing Director",
-  },
-];
+// const testimonials = [
+//   {
+//     program: "Platinum",
+//     description:
+//       "Being part of BITO means being part of Bihar’s growth and transformation. It’s about honoring our roots, and shaping a future. I’m grateful to contribute to this meaningful journey.",
+//     image: new URL("../assets/testimonials/t1.jpg", import.meta.url),
+//     name: "Farooque",
+//     title: "Customer Success Manager, Dubai",
+//   },
+//   {
+//     program: "Diamond",
+//     description:
+//       "BITO is more than an organization—it’s a movement. By empowering entrepreneurs, fostering innovation, and honoring Bihar’s rich heritage, it’s building bridges between our past and our global future. I’m proud to be part of this journey toward a more united and prosperous Bihar.",
+//     image: new URL("../assets/testimonials/t2.jpg", import.meta.url),
+//     name: "Upendra",
+//     // title: "Product lead at EducationWave",
+//   },
+//   {
+//     program: "Silver",
+//     description:
+//       "As an education leader, I admire BITO’s focus on experiential learning through trade exposure, internships, and global forums. These initiatives broaden horizons and make learning truly meaningful.",
+//     image: new URL("../assets/testimonials/t3.jpg", import.meta.url),
+//     name: "Abhilasha Singh",
+//     title: "Principal Shining Star International School, Abu Dhabi",
+//   },
+//   {
+//     program: "Gold",
+//     description:
+//       "We are expert in delivering exhibitions, interior and infrastructure services across worldwide having company presence in Patna, Dubai, Milano, Warsaw. Now just we launched in Patna Indl Area also.",
+//     image: new URL("../assets/testimonials/t4.jpg", import.meta.url),
+//     name: "Mukesh Pandey",
+//     title: "Managing Director",
+//   },
+// ];
 
 function Testimonials() {
   return (
